swagger 및 rest API 명세 2차수정
